+++ packages/v3/src/main.ts
@@ -64,19 +64,6 @@ declare global {
   }
 }
 
-// TODO: this should be removed if all works well
-/**
- * HACK: Cluster should be loaded conditionally
- * However in the web version, it's not possible to write
- * `import 'gmap-vue/src/components/cluster'`, so we need to
- * import it anyway (but we don't have to register it)
- * Therefore we use babel-plugin-transform-inline-environment-variables to
- * set BUILD_DEV to truthy / falsy
- */
-// const Cluster = ((s) => s.default || s)(
-//   require('./components/cluster-icon.vue')
-// );
-
 /**
  * @var
  * @type {Object|undefined}